Fixed out-of-date check_track.py and moved query_drone_poses and query_track_poses into envs/utils.py
