+++ lsy_drone_racing/envs/utils.py
@@ -8,6 +8,7 @@ from typing import TYPE_CHECKING
 import jax
 import jax.numpy as jnp
 import numpy as np
+from drone_estimators.ros_nodes.ros2_connector import ROSConnector
 from jax.numpy import vectorize
 from jax.scipy.spatial.transform import Rotation as JR
 from ml_collections import ConfigDict
@@ -15,6 +16,7 @@ from scipy.spatial.transform import Rotation as R
 
 if TYPE_CHECKING:
     from jax import Array
+    from numpy.typing import NDArray
 
 
 def load_track(track: ConfigDict) -> tuple[ConfigDict, ConfigDict, ConfigDict]:
@@ -51,6 +53,74 @@ def load_track(track: ConfigDict) -> tuple[ConfigDict, ConfigDict, ConfigDict]:
     }
     drones["quat"] = R.from_euler("xyz", drones["rpy"]).as_quat().astype(np.float32)
     return ConfigDict(gates), ConfigDict(obstacles), ConfigDict(drones)
+
+
+def query_track_poses(n_gates: int, n_obstacles: int) -> tuple[NDArray, NDArray, NDArray]:
+    """Query the track poses from the motion capture system.
+
+    Args:
+        n_gates: Number of gates in the track.
+        n_obstacles: Number of obstacles in the track.
+
+    Returns:
+        A tuple containing:
+            - gate_pos: An (n_gates, 3) array of gate positions.
+            - gate_quat: An (n_gates, 4) array of gate orientations as quaternions.
+            - obstacle_pos: An (n_obstacles, 3) array of obstacle positions.
+    """
+    gate_pos = np.zeros((n_gates, 3), dtype=np.float32)
+    gate_quat = np.zeros((n_gates, 4), dtype=np.float32)
+    obstacle_pos = np.zeros((n_obstacles, 3), dtype=np.float32)
+
+    tf_names = [f"gate{i}" for i in range(1, n_gates + 1)]
+    tf_names += [f"obstacle{i}" for i in range(1, n_obstacles + 1)]
+    try:
+        ros_connector = ROSConnector(tf_names=tf_names, timeout=10.0)
+        pos, quat = ros_connector.pos, ros_connector.quat
+    finally:
+        ros_connector.close()
+    try:
+        for i in range(n_gates):
+            gate_pos[i, ...] = pos[f"gate{i + 1}"]
+            gate_quat[i, ...] = quat[f"gate{i + 1}"]
+        for i in range(n_obstacles):
+            obstacle_pos[i, ...] = pos[f"obstacle{i + 1}"]
+    except KeyError as e:
+        raise ValueError(
+            f"Could not find all track objects in the ROS TF tree: {e}. \
+                        Have you enabled the track objects in Vicon/ \
+                        started the motion capture tracking node?"
+        ) from e
+    return gate_pos, gate_quat, obstacle_pos
+
+
+def query_drone_poses(drone_names: list[str]) -> tuple[NDArray, NDArray]:
+    """Query the drone positions from the motion capture system and estimator node.
+
+    Args:
+        drone_names: List of drone estimator names.
+
+    Returns:
+        A tuple containing:
+            - drone_pos: An (n_drones, 3) array of drone positions.
+            - drone_quat: An (n_drones, 4) array of drone orientations as quaternions.
+    """
+    drone_pos = np.zeros((len(drone_names), 3), dtype=np.float32)
+    drone_quat = np.zeros((len(drone_names), 4), dtype=np.float32)
+    try:
+        ros_connector = ROSConnector(estimator_names=drone_names, timeout=10.0)
+        for i, drone_name in enumerate(drone_names):
+            drone_pos[i, ...] = ros_connector.pos[drone_name]
+            drone_quat[i, ...] = ros_connector.quat[drone_name]
+    except KeyError as e:
+        raise ValueError(
+            f"Could not find drone in the ROS estimator messages: {e}. \
+                        Have you enabled the drone in Vicon/ \
+                        started the estimator node?"
+        ) from e
+    finally:
+        ros_connector.close()
+    return drone_pos, drone_quat
 
 
 @jax.jit
