+++ scripts/check_track.py
@@ -7,6 +7,7 @@ import fire
 import rclpy
 from scipy.spatial.transform import Rotation as R
 
+from lsy_drone_racing.envs.utils import query_track_poses
 from lsy_drone_racing.utils import load_config
 from lsy_drone_racing.utils.checks import check_race_track
 
@@ -21,10 +22,25 @@ def main(config: str = "level2.toml"):
     """
     rclpy.init()
     config = load_config(Path(__file__).resolve().parents[1] / "config" / config)
-    gates_pos = [g["pos"] for g in config.env.track.gates]
-    gates_quat = [R.from_euler("xyz", g["rpy"]).as_quat() for g in config.env.track.gates]
-    obstacle_pos = [o["pos"] for o in config.env.track.obstacles]
-    check_race_track(gates_pos, gates_quat, obstacle_pos, rng_config=config.env.randomizations)
+    nominal_gates_pos = [g["pos"] for g in config.env.track.gates]
+    nominal_gates_quat = [R.from_euler("xyz", g["rpy"]).as_quat() for g in config.env.track.gates]
+    nominal_obstacles_pos = [o["pos"] for o in config.env.track.obstacles]
+    n_gates = len(nominal_gates_pos)
+    n_obstacles = len(nominal_obstacles_pos)
+    gates, obstacles = dict(), dict()
+    gates["pos"], gates["quat"], obstacles["pos"] = query_track_poses(
+        n_gates=n_gates, n_obstacles=n_obstacles
+    )
+    check_race_track(
+        gates_pos=gates["pos"],
+        nominal_gates_pos=nominal_gates_pos,
+        gates_quat=gates["quat"],
+        nominal_gates_quat=nominal_gates_quat,
+        obstacles_pos=obstacles["pos"],
+        nominal_obstacles_pos=nominal_obstacles_pos,
+        rng_config=config.env.randomizations,
+    )
+
     logger.info("Race track check passed")
