+++ scripts/save_track_as_config.py
@@ -4,14 +4,12 @@ import logging
 from pathlib import Path
 
 import fire
-import numpy as np
 import rclpy
 import toml
-from drone_estimators.ros_nodes.ros2_connector import ROSConnector
 from ml_collections import ConfigDict
-from numpy.typing import NDArray
 from scipy.spatial.transform import Rotation as R
 
+from lsy_drone_racing.envs.utils import query_drone_poses, query_track_poses
 from lsy_drone_racing.utils import load_config
 
 logger = logging.getLogger("rosout." + __name__)
@@ -48,74 +46,6 @@ def main(config: str = "level2.toml", save_config_to: str = "real_track.toml"):
 if __name__ == "__main__":
     logging.basicConfig(level=logging.INFO)
     fire.Fire(main)
-
-
-def query_track_poses(n_gates: int, n_obstacles: int) -> tuple[NDArray, NDArray, NDArray]:
-    """Query the track poses from the motion capture system.
-
-    Args:
-        n_gates: Number of gates in the track.
-        n_obstacles: Number of obstacles in the track.
-
-    Returns:
-        A tuple containing:
-            - gate_pos: An (n_gates, 3) array of gate positions.
-            - gate_quat: An (n_gates, 4) array of gate orientations as quaternions.
-            - obstacle_pos: An (n_obstacles, 3) array of obstacle positions.
-    """
-    gate_pos = np.zeros((n_gates, 3), dtype=np.float32)
-    gate_quat = np.zeros((n_gates, 4), dtype=np.float32)
-    obstacle_pos = np.zeros((n_obstacles, 3), dtype=np.float32)
-
-    tf_names = [f"gate{i}" for i in range(1, n_gates + 1)]
-    tf_names += [f"obstacle{i}" for i in range(1, n_obstacles + 1)]
-    try:
-        ros_connector = ROSConnector(tf_names=tf_names, timeout=10.0)
-        pos, quat = ros_connector.pos, ros_connector.quat
-    finally:
-        ros_connector.close()
-    try:
-        for i in range(n_gates):
-            gate_pos[i, ...] = pos[f"gate{i + 1}"]
-            gate_quat[i, ...] = quat[f"gate{i + 1}"]
-        for i in range(n_obstacles):
-            obstacle_pos[i, ...] = pos[f"obstacle{i + 1}"]
-    except KeyError as e:
-        raise ValueError(
-            f"Could not find all track objects in the ROS TF tree: {e}. \
-                        Have you enabled the track objects in Vicon/ \
-                        started the motion capture tracking node?"
-        ) from e
-    return gate_pos, gate_quat, obstacle_pos
-
-
-def query_drone_poses(drone_names: list[str]) -> tuple[NDArray, NDArray]:
-    """Query the drone positions from the motion capture system and estimator node.
-
-    Args:
-        drone_names: List of drone estimator names.
-
-    Returns:
-        A tuple containing:
-            - drone_pos: An (n_drones, 3) array of drone positions.
-            - drone_quat: An (n_drones, 4) array of drone orientations as quaternions.
-    """
-    drone_pos = np.zeros((len(drone_names), 3), dtype=np.float32)
-    drone_quat = np.zeros((len(drone_names), 4), dtype=np.float32)
-    try:
-        ros_connector = ROSConnector(estimator_names=drone_names, timeout=10.0)
-        for i, drone_name in enumerate(drone_names):
-            drone_pos[i, ...] = ros_connector.pos[drone_name]
-            drone_quat[i, ...] = ros_connector.quat[drone_name]
-    except KeyError as e:
-        raise ValueError(
-            f"Could not find drone in the ROS estimator messages: {e}. \
-                        Have you enabled the drone in Vicon/ \
-                        started the estimator node?"
-        ) from e
-    finally:
-        ros_connector.close()
-    return drone_pos, drone_quat
 
 
 def update_level_config(
